seperate providers to meaningful subdirs for upcoming addition, and adds doc comments and breadcrumbs
